docs: :memo: added github pages
implment github ages with docsify

Closes #13
