+++ README.md
@@ -1,37 +1,3 @@
-# Typescript Design Patterns
+# Examples for Dessign Patterns
 
-[Español](README-es.md)
-
-Examples for Dessign Patterns in TypeScript
-
-## Creational Design Paterns
-
-1. Factory method
-1. Abstract Factory
-1. Builder
-1. [Prototype](./designs/creationalPatterns/prototype/README.md)
-1. Singleton
-
-## Structural Design Patterns
-
-1. [Adapter](designs/structuralPatterns/adapter/README.md)
-1. Bridge
-1. Composite
-1. Decorator
-1. Facede
-1. Flyweight
-1. Proxy
-
-## Behavioral Design Patterns
-
-1. Chain of responsibility
-1. Command
-1. Interpreter
-1. Iterator
-1. Mediator
-1. Memento
-1. [Observer](designs/behavioralPatterns/observer/README.md)
-1. State
-1. Strategy
-1. Template method
-1. Visitor
+[Go to Documentation page] (https://jeresoftx.github.io/dessign-patterns)
