--- designs/behavioralPatterns/observer/README-es.md
@@ -1,10 +1,0 @@
-# Patrón de diseño observador
-
-[Incio](../../../README-es.md)
-
-Define una relación uno a muchos entre objetos, de modo que cuando un objeto cambia de estado, todos sus observadores son notificados y actualizados automáticamente. Ejemplo: Implementar notificaciones en una aplicación en tiempo real.
-
-## Ejemplos
-
-1. [Typescript](./observer.ts)
-1. [Python](./observer.py)
