--- designs/behavioralPatterns/observer/README.md
@@ -1,10 +1,0 @@
-# Dessing Pattern Observator
-
-[Home](../../../README.md)
-
-Define a one-to-many relationship between objects, so that when an object changes its state, all its observers are notified and updated automatically. Example: Implement notifications in a real-time application.
-
-## Examples
-
-1. [Typescript](./observer.ts)
-1. [Python](./observer.py)
