--- designs/structuralPatterns/adapter/README-es.md
@@ -1,11 +1,0 @@
-# Patrón de diseño adapter
-
-[Incio](../../../README-es.md)
-
-El patrón de diseño Adaptador (Adapter Pattern) se utiliza para permitir que dos interfaces incompatibles trabajen juntas. Un adaptador actúa como un intermediario entre dos objetos que no pueden comunicarse directamente.
-
-## Ejemplos
-
-1. [Typescript](./adapter.ts)
-1. [Python](./adapter.py)
-1. [Kotlin](./adapter.kt)
