--- designs/structuralPatterns/adapter/README.md
@@ -1,11 +1,0 @@
-# Dessing Pattern Adapter
-
-[Incio](../../../README-es.md)
-
-The Adapter Pattern is used to allow two incompatible interfaces to work together. An adapter acts as an intermediary between two objects that cannot communicate directly.
-
-## Examples
-
-1. [Typescript](./adapter.ts)
-1. [Python](./adapter.py)
-1. [Kotlin](./adapter.kt)
