+++ docs/behavioralPatterns/behavioralPatterns.md
@@ -1,0 +1,13 @@
+# Behavioral Design Patterns
+
+1. Chain of responsibility
+1. Command
+1. Interpreter
+1. Iterator
+1. Mediator
+1. Memento
+1. [Observer](/behavioralPatterns/observer/index.md)
+1. State
+1. Strategy
+1. Template method
+1. Visitor
