+++ docs/behavioralPatterns/observer.md
@@ -1,0 +1,19 @@
+# Dessing Pattern Observator
+
+## Description
+
+Define a one-to-many relationship between objects, so that when an object changes its state, all its observers are notified and updated automatically. Example: Implement notifications in a real-time application.
+
+## Example code
+
+<!-- tabs:start -->
+
+#### **Typescript**
+
+[observer.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/5a05e359d608a22d16851d500d76335aab2407be/designs/behavioralPatterns/observer/observer.ts ':include :type=code')
+
+#### **Python**
+
+[observer.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/5a05e359d608a22d16851d500d76335aab2407be/designs/behavioralPatterns/observer/observer.py ':include :type=code')
+
+<!-- tabs:end -->
