+++ docs/es-mx/README.md
@@ -1,0 +1,41 @@
+# Patrones de diseño
+
+Ejemplo de patrones de diseño
+
+## Patrones de Diseño Creacionales
+
+Los patrones de diseño creacionales son un conjunto de patrones en el diseño de software que se utilizan para crear objetos de manera más flexible y eficiente. Aquí hay algunos ejemplos de patrones de diseño creacionales:
+
+1. Abstract Factory (Fábrica Abstracta)
+1. Builder (Constructor)
+1. Factory method (Método de Fábrica)
+1. [Prototype (prototipo)](./designs/creationalPatterns/prototype/README-es.md)
+1. Singleton (singleton)
+
+## Patrones de Diseño Estructurales
+
+Los patrones de diseño estructurales son un conjunto de patrones en el diseño de software que se enfocan en cómo los objetos y clases se componen para formar estructuras más grandes y complejas. Aquí tienes algunos ejemplos de patrones de diseño estructurales:
+
+1. [Adapter](designs/structuralPatterns/adataper/README-es.md)
+1. Bridge
+1. Composite
+1. Decorator
+1. Facede
+1. Flyweight
+1. Proxy
+
+## Patrones de Diseño del comportamiento
+
+Los patrones de diseño del comportamiento son un conjunto de patrones en el diseño de software que se centran en cómo los objetos interactúan y comunican entre sí. Aquí hay algunos ejemplos de patrones de diseño del comportamiento:
+
+1. Chain of responsibility
+1. Command
+1. Interpreter
+1. Iterator
+1. Mediator
+1. Memento
+1. [Observer (observador)](designs/behavioralPatterns/observer/README-es.md)
+1. State
+1. Strategy
+1. Template method
+1. Visitor
