+++ docs/es-mx/behavioralPatterns/observer.md
@@ -1,0 +1,19 @@
+# Patrón de diseño observador
+
+[Incio](../../../README-es.md)
+
+Define una relación uno a muchos entre objetos, de modo que cuando un objeto cambia de estado, todos sus observadores son notificados y actualizados automáticamente. Ejemplo: Implementar notificaciones en una aplicación en tiempo real.
+
+## Ejemplos de código
+
+<!-- tabs:start -->
+
+#### **Typescript**
+
+[observer.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/5a05e359d608a22d16851d500d76335aab2407be/designs/behavioralPatterns/observer/observer.ts ':include :type=code')
+
+#### **Python**
+
+[observer.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/5a05e359d608a22d16851d500d76335aab2407be/designs/behavioralPatterns/observer/observer.py ':include :type=code')
+
+<!-- tabs:end -->
