+++ docs/es-mx/creationalPatterns/creationalPatterns.md
@@ -1,0 +1,11 @@
+# Patrones de diseño creacionales
+
+## Definición
+
+## Ejemplos
+
+1. Factory method
+1. Abstract Factory
+1. Builder
+1. [Prototype](/creationalPatterns/prototype/index.md 'Patrón de diseño prototipo')
+1. Singleton
