+++ docs/es-mx/structuralPatterns/adapter.md
@@ -1,0 +1,23 @@
+# Patrón de diseño adaptador
+
+## Descripción
+
+El patrón de diseño Adaptador (Adapter Pattern) se utiliza para permitir que dos interfaces incompatibles trabajen juntas. Un adaptador actúa como un intermediario entre dos objetos que no pueden comunicarse directamente.
+
+## Ejemplos
+
+<!-- tabs:start -->
+
+#### **Typescript**
+
+[prototype.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/main/designs/structuralPatterns/adapter/adapter.ts ':include :type=code')
+
+#### **Python**
+
+[prototype.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/main/designs/structuralPatterns/adapter/adapter.py ':include :type=code')
+
+#### **Kotlin**
+
+[prototype.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/main/designs/structuralPatterns/adapter/adapter.kt ':include :type=code')
+
+<!-- tabs:end -->
