+++ docs/es-mx/structuralPatterns/structuralPatterns.md
@@ -1,0 +1,9 @@
+# Structural Design Patterns
+
+1. [Adapter](/structuralPatterns/adapter/README.md)
+1. Bridge
+1. Composite
+1. Decorator
+1. Facede
+1. Flyweight
+1. Proxy
