+++ docs/structuralPatterns/adapter.md
@@ -1,0 +1,21 @@
+# Dessing Pattern PrAdapterototype
+
+The Adapter Pattern is used to allow two incompatible interfaces to work together. An adapter acts as an intermediary between two objects that cannot communicate directly.
+
+## Code Examples
+
+<!-- tabs:start -->
+
+#### **Typescript**
+
+[prototype.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/main/designs/structuralPatterns/adapter/adapter.ts ':include :type=code')
+
+#### **Python**
+
+[prototype.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/main/designs/structuralPatterns/adapter/adapter.py ':include :type=code')
+
+#### **Kotlin**
+
+[prototype.py](https://raw.githubusercontent.com/jeresoftx/design-patterns/main/designs/structuralPatterns/adapter/adapter.kt ':include :type=code')
+
+<!-- tabs:end -->
